requirements.txt
python-dotenv
fastapi
uvicorn
streamlit
pandas
numpy
scikit-learn
xgboost
shap
pydantic
openai
google-generativeai
joblib
matplotlib
seaborn
requests
plotly
